Edit functions

+++ addon/components/jsplumb-node.js
@@ -12,8 +12,6 @@ import layout from '../templates/components/jsplumb-node';
 
 import { getProperties, setProperties } from '@ember/object';
 
-const noop = () => {};
-
 export default Component.extend(ChildMixin, {
   layout,
 
@@ -103,8 +101,8 @@ export default Component.extend(ChildMixin, {
 
       jsPlumb.revalidate(element);
 
-      const onResize = this.parentComponent.parentComponent.onResizeNode || noop;
-      onResize(node, element);
+      const onResize = this.parentComponent.parentComponent.onResizeNode;
+      onResize && onResize(node, element);
     },
     editNode() {
       const node = this.get('node');
@@ -122,8 +120,8 @@ export default Component.extend(ChildMixin, {
         )
       );
 
-      const onEdit = this.parentComponent.parentComponent.onEditNode || noop;
-      onEdit(node, element, { edges: nodeEdges, connections, definition });
+      const onEdit = this.parentComponent.parentComponent.onEditNode;
+      onEdit && onEdit(node, element, { edges: nodeEdges, connections, definition });
     },
     deleteNode() {
       const node = this.get('node');
@@ -136,8 +134,8 @@ export default Component.extend(ChildMixin, {
       jsPlumb.remove(element);
       jsplumbUtils.removeNode(node, nodes, edges);
 
-      const onRemove = this.parentComponent.parentComponent.onRemoveNode || noop;
-      onRemove(node, element);
+      const onRemove = this.parentComponent.parentComponent.onRemoveNode;
+      onRemove && onRemove(node, element);
     }
   }
 });
